WIP: Simulation bug attempted fix.
This fix came from a bug in the simulation that does not allow the
telescope to go past 0/360 degrees (in either direction). The attempted
fix is to split movements up into two: Move to zero, and then move to
the target. This does not work, however. More investigation will need to
be done.

Issue #252